+++ scripts/run-jazzy.sh
@@ -1,4 +1,4 @@
 #!/bin/bash
 
-jazzy --clean --source-directory src  --output docs -a Luis Padron -u https://luispadron.com -m UICircularProgressRing --module-version 1.8.5 -g https://github.com/luispadron/UICircularProgressRing --theme apple
+jazzy --clean --source-directory src  --output docs -a Luis Padron -u https://luispadron.com -m UICircularProgressRing --module-version 2.0.0 -g https://github.com/luispadron/UICircularProgressRing --theme apple